backend/requirements.txt
# ========================
# NEXUS AI - Backend Dependencies
# ========================

# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1

# Database
sqlalchemy==2.0.35
alembic==1.13.3
aiosqlite==0.20.0

# Vector Database
chromadb==0.5.7

# AI/ML
ollama==0.3.3
langchain==0.3.1
langchain-community==0.3.1
langchain-core==0.3.6
transformers==4.45.1
torch==2.4.1
sentence-transformers==3.1.1
scikit-learn==1.5.2
numpy==1.26.4
pandas==2.2.3

# Voice
TTS==0.22.0
openai-whisper==20231117
sounddevice==0.5.0
soundfile==0.12.1
pyaudio==0.2.14
webrtcvad==2.0.10

# Email & Communication
aiosmtplib==3.0.2
python-multipart==0.0.12
aiohttp==3.10.8
beautifulsoup4==4.12.3

# Report Generation
reportlab==4.2.4
openpyxl==3.1.5
markdown==3.7
jinja2==3.1.4
weasyprint==62.3

# Scheduling
apscheduler==3.10.4
celery==5.4.0

# System Monitoring
psutil==6.0.0
GPUtil==1.4.0

# Home Automation
paho-mqtt==2.1.0
websockets==13.1

# Security
cryptography==43.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Utilities
httpx==0.27.2
loguru==0.7.2
rich==13.9.1
tenacity==9.0.0
cachetools==5.5.0
watchdog==5.0.2
python-dateutil==2.9.0
Pillow==10.4.0
